# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreatePrendaDTO:
    properties:
      descripcion:
        type: string
      id_genero:
        type: integer
      id_temporada:
        type: integer
      id_tipo_prenda:
        type: integer
      maternal:
        type: boolean
      nombre_prenda:
        type: string
      tallas_disponibles:
        type: string
      url_imagen:
        type: string
    required:
    - id_genero
    - id_temporada
    - id_tipo_prenda
    - nombre_prenda
    - tallas_disponibles
    type: object
...
  models.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
    type: object
...
  models.SearchPrendasDTO:
    properties:
      generos:
        items:
          type: integer
        type: array
      limit:
        type: integer
      maternal:
        type: boolean
      nombre:
        type: string
      page:
        type: integer
      proveedores:
        items:
          type: integer
        type: array
      tallas:
        items:
          type: string
        type: array
      temporadas:
        items:
          type: integer
        type: array
      tipos_prenda:
        items:
          type: integer
        type: array
    type: object
  models.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorInfo'
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  models.UpdatePrendaDTO:
    properties:
      descripcion:
        type: string
      id_genero:
        type: integer
      id_temporada:
        type: integer
      id_tipo_prenda:
        type: integer
      maternal:
        type: boolean
      nombre_prenda:
        type: string
      tallas_disponibles:
        type: string
      url_imagen:
        type: string
    type: object
...
  models.VincularProveedorDTO:
    properties:
      id_empresa:
        type: integer
    required:
    - id_empresa
    type: object
host: localhost:8084
info:
  contact:
    email: soporte@sistal.com
    name: Soporte API
  description: |-
    API del Microservicio de Catálogo de Prendas y Uniformes
    Gestiona el catálogo de prendas, uniformes, temporadas y datos maestros
  license:
    name: MIT
  title: MS-Catalogo API
  version: "1.0"
paths:
...
  /api/v1/prendas:
    get:
      consumes:
      - application/json
      description: Obtiene lista paginada de prendas con filtros
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por tipo de prenda
        in: query
        name: id_tipo_prenda
        type: integer
      - description: Filtrar por género
        in: query
        name: id_genero
        type: integer
      - description: Filtrar por temporada
        in: query
        name: id_temporada
        type: integer
      - description: Solo prendas maternales
        in: query
        name: maternal
        type: boolean
      - description: Buscar en nombre o descripción
        in: query
        name: search
        type: string
      - default: '"nombre_prenda"'
        description: Campo para ordenar
        in: query
        name: sort_by
        type: string
      - default: '"asc"'
        description: Orden (asc|desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Listar Prendas
      tags:
      - Prendas
    post:
      consumes:
      - application/json
      description: Crea una nueva prenda en el catálogo
      parameters:
      - description: Datos de la prenda
        in: body
        name: prenda
        required: true
        schema:
          $ref: '#/definitions/models.CreatePrendaDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Crear Prenda
      tags:
      - Prendas
  /api/v1/prendas/{id_prenda}:
    delete:
      consumes:
      - application/json
      description: Elimina una prenda del catálogo
      parameters:
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Eliminar Prenda
      tags:
      - Prendas
    get:
      consumes:
      - application/json
      description: Obtiene información detallada de una prenda
      parameters:
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Obtener Prenda por ID
      tags:
      - Prendas
    put:
      consumes:
      - application/json
      description: Actualiza una prenda existente
      parameters:
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: prenda
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePrendaDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Actualizar Prenda
      tags:
      - Prendas
  /api/v1/prendas/{id_prenda}/proveedores:
    post:
      consumes:
      - application/json
      description: Asocia un proveedor a una prenda
      parameters:
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      - description: ID del proveedor
        in: body
        name: proveedor
        required: true
        schema:
          $ref: '#/definitions/models.VincularProveedorDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Vincular Proveedor a Prenda
      tags:
      - Prendas
  /api/v1/prendas/buscar:
    post:
      consumes:
      - application/json
      description: Búsqueda avanzada con múltiples filtros
      parameters:
      - description: Criterios de búsqueda
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/models.SearchPrendasDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Búsqueda Avanzada de Prendas
      tags:
      - Prendas
...
schemes:
- http
- https
swagger: "2.0"
...